+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Unreleased
 Added
 -----
 
+- New algorithm for long-range corrected Hamiltonian
+
 
 Changed
 -------